+++ src/main/groovy/ru/endlesscode/bukkitgradle/Dependencies.groovy
@@ -1,4 +1,4 @@
-package ru.endlesscode.bukkitgradle.util
+package ru.endlesscode.bukkitgradle
 
 import org.gradle.api.Project
 import org.gradle.api.artifacts.Dependency
