+++ src/main/groovy/ru/endlesscode/bukkitgradle/server/idea/IdeaRunConfigurationBuilder.groovy
@@ -1,7 +1,7 @@
 package ru.endlesscode.bukkitgradle.server.idea
 
 import groovy.xml.MarkupBuilder
-import ru.endlesscode.bukkitgradle.extension.RunConfiguration
+import ru.endlesscode.bukkitgradle.server.extension.RunConfiguration
 import ru.endlesscode.bukkitgradle.server.ServerConstants
 
 import java.nio.file.Files
