+++ src/main/groovy/ru/endlesscode/bukkitgradle/server/task/RunServer.groovy
@@ -1,9 +1,9 @@
-package ru.endlesscode.bukkitgradle.task
+package ru.endlesscode.bukkitgradle.server.task
 
 import org.gradle.api.DefaultTask
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.TaskAction
-import ru.endlesscode.bukkitgradle.extension.RunConfiguration
+import ru.endlesscode.bukkitgradle.server.extension.RunConfiguration
 import ru.endlesscode.bukkitgradle.server.ServerCore
 import ru.endlesscode.bukkitgradle.server.script.RunningScript
